Conflicts:

	src/mesa/main/context.c

+++ progs/tools/trace/gltrace_support.cc
@@ -20,6 +20,7 @@
  */
 
 #include "gltrace_support.h"
+#include <assert.h>
 #include <sstream>
 #include <fstream>
 #include <iomanip>
@@ -34,7 +35,7 @@ namespace {
   {
     static char buf[4096];
     int status;
-    unsigned int length = sizeof(buf)-1;
+    size_t length = sizeof(buf)-1;
     
     memset (buf, 0, sizeof(buf));
